refactor(UserContext): Move UserContext to specific file
