+++ front/package.json
@@ -10,7 +10,9 @@
     },
     "dependencies": {
         "@types/js-cookie": "^3.0.2",
+        "@types/jwt-decode": "^3.1.0",
         "js-cookie": "^3.0.1",
+        "jwt-decode": "^3.1.2",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-router": "^6.6.1",
