+++ front/src/App.tsx
@@ -1,34 +1,13 @@
-import { useState } from 'react';
 import AppRouter from './components/router/AppRouter';
-import { UserContext } from './context/UserContext';
+import UserContextProvider from './context/UserContext';
 
 function App() {
-    const [logged, setLogged] = useState(false);
-    const [access_token, setAccessToken] = useState('');
-
-    function updateUser() {}
-
-    function getAccessToken() {}
-
     return (
-        <UserContext.Provider
-            value={{
-                auth: {
-                    logged,
-                    setLogged,
-                    access_token,
-                    setAccessToken,
-                },
-                updateUser,
-                getAccessToken,
-                id: -1,
-                name: '',
-            }}
-        >
+        <UserContextProvider>
             <div className="App">
                 <AppRouter />
             </div>
-        </UserContext.Provider>
+        </UserContextProvider>
     );
 }
